Transition from raw `INVALID_REF` usage to strong typed reference management
- Replaced raw `INVALID_REF` constants with strongly typed `.isValid()` and `.isInvalid()` checks across all parser modules.
- Introduced `StrongRef` template in `Core/Types.h` to encapsulate reference handling with type safety.
- Updated `AstNodeRef`, `TokenRef`, and other references to use `StrongRef` for improved code reliability.
- Refactored multiple parsing methods to align with the new reference handling paradigm, adhering to cleaner and safer coding practices.

+++ src/Main/FileManager.cpp
@@ -17,7 +17,7 @@ FileRef FileManager::addFile(fs::path path)
     if (it != paths_.end())
         return it->second;
 
-    const auto fileRef = static_cast<FileRef>(files_.size() + 1);
+    const auto fileRef = static_cast<FileRef>(static_cast<uint32_t>(files_.size() + 1));
     paths_[path]       = fileRef;
 
     files_.emplace_back(std::make_unique<SourceFile>(std::move(path)));
